Refactor consensus && miner solver
1. change nonce u64 to u32
2. change mining_hash from hashvalue to vec mining_blob
3. refactor solver trait for support dynamic link

+++ vm/types/src/genesis_config.rs
@@ -667,7 +667,7 @@ pub struct GenesisConfig {
     /// Genesis difficulty, should match consensus in different ChainNetwork.
     pub difficulty: U256,
     /// Genesis consensus nonce.
-    pub nonce: u64,
+    pub nonce: u32,
     /// Pre mine STC amount to Association account.
     pub pre_mine_amount: u128,
     /// If time_mint_amount >0, Issue a LinearTimeMintKey to Association account
